+++ application/src/File/Thumbnailer/Imagick.php
@@ -7,7 +7,7 @@ use Omeka\File\Exception;
 use Omeka\File\Manager as FileManager;
 use Omeka\File\TempFileFactory;
 
-class ImagickThumbnailer extends AbstractThumbnailer
+class Imagick extends AbstractThumbnailer
 {
     /**
      * @var TempFileFactory
