+++ application/src/Service/File/Thumbnailer/ImageMagickFactory.php
@@ -1,8 +1,8 @@
 <?php
-namespace Omeka\Service\FileThumbnailer;
+namespace Omeka\Service\File\Thumbnailer;
 
 use Interop\Container\ContainerInterface;
-use Omeka\File\Thumbnailer\ImageMagickThumbnailer;
+use Omeka\File\Thumbnailer\ImageMagick;
 use Zend\ServiceManager\Factory\FactoryInterface;
 
 class ImageMagickFactory implements FactoryInterface
@@ -10,11 +10,11 @@ class ImageMagickFactory implements FactoryInterface
     /**
      * Create the ImageMagick thumbnailer service.
      *
-     * @return ImageMagickThumbnailer
+     * @return ImageMagick
      */
     public function __invoke(ContainerInterface $services, $requestedName, array $options = null)
     {
-        return new ImageMagickThumbnailer(
+        return new ImageMagick(
             $services->get('Omeka\Cli'),
             $services->get('Omeka\File\TempFileFactory')
         );
